New (live) ebuild games-action/oolite: Modern clone of the classic Elite
Oolite is a modern clone of the classic space trading/combat simulator
Elite built using ObjC/Cocoa (originally written for OS X) and OpenGL.

It is highly expandable with hundreds of expansions available.  This
version is better than the one in portage because it compiles with
modern versions of GCC, and uses OBJCFLAGS optimization flags if set!

 ;-)