New ebuild dev-libs/boehm-gc
This uses prune_libtool_files instead of the find command which wasn't
doing it's job.

I noticed since gcc links to boehm-gc when build objc support and if the
.la files are present -fopenmp gets added to the inherited_linker_flags
if it's in the flags of a dependency from autopar even if OpenMP isn't
required and has been dropped by as-needed, but during building objc
libgomp isn't available to xgcc.